fix: setupBinary() promise resolves when complete
Right now there is an edge case where setupBinary() can resolve before a
download is done: if called while a download for that version is already
running, setupBinary() returns. When this happens, the caller may find
their RunnableVersion does not have the state that was expected.

This PR keeps track of pending downloads so that when a duplicate call
occurs, all callers get a Promise that settles when the download does.

This is similar to how renderer/content safeguards against parallel
downloads of the same zipfile:
https://github.com/electron/fiddle/blob/858e4b1834832e0ea5ccf64825c2bd7d201a8a2b/src/renderer/content.ts#L57