perf: faster isReleasedMajor()
Just a small one-liner. I wrote the previous version of this last week
and keep seeing this line have an outsized sample count in the coverage
tests. To get the max item from the versions array we don't need to sort
the array; we can walk through it with reduce().

Brings this line from O(n log n) to O(n).