feat: add an icon for Electron DLs being unzipped
Right now there's a visual wart that during an Electron download the
icon goes from 'cloud' (unknown) -> 'cloud-download' (downloading) ->
'cloud' (unzipping) -> 'saved' (saved). That second 'cloud' is a little
confusing because for a moment it looks like the download disappeared.

This PR uses 'compressed' as the unzip icon to avoid that mixed signal.