feat: make document.title track application state
In master, the title is only updated when `replaceFiddle()` is called.

This PR makes title a computed appState property that updates whenever
[ gistId, isUnsaved localPath, templateName ] changes. App listens for
this to keep document.title in sync with state.title.

This change also has a nice side-effect that loading `Show Me` templates
or pressing `New Test` will show the template name.

Fixes #599.