refactor: remove extract-zip from top dependencies
decompress and extract-zip are very similar; the reason for choosing
one over the other is renderer/content.ts needs to be able to operate
on an input stream instead of an input file, so decompress wins.

In practice this doesn't change anything *yet* since both are still
required in our dependency tree [electron-rebuild -> got -> decompress]
and [electron-repackager -> extract-zip], but at least this way it's
not blocked by anything in Electron Fiddle.