feat: replace 'unsupported' filter with 'obsolete'
Back in 2018 the project made an Electron release tagged 'unsupported'
because of reasons. It was a one-off thing that isn't done anymore and
we don't need to have a filter for it, so this PR removes that filter.

More useful is the question "is this version of Electron End-of-Lifed?"
because Fiddle should encourage users to write their code on supported
versions of Electron. So this PR also *adds* a filter called "Obsolete"
that lets uses toggle display of end-of-lifed versions.