feat: new version filters in File > Preferences > Electron (#656)
* feat: simplify states to show in electron settings

In master, users can filter versions by VersionState. This introduces
some options without any use case, e.g. filtering based on whether or
not a version is currently being unzipped (an operation that takes less
than a second).

This change reduces those options down to the one that matters:
"Show undownloaded versions?"

* feat: replace 'unsupported' filter with 'obsolete'

Back in 2019(?) the project made a Electron release tagged 'unsupported'
because of reasons. It was a one-off thing that isn't done anymore and
we don't need to have a filter for it, so this PR removes that filter.

More useful is the question "is this version of Electron End-of-Lifed?"
because Fiddle should encourage users to write their code on supported
versions of Electron. So this PR also *adds* a filter called "Obsolete"
that lets uses toggle display of end-of-lifed versions.