refactor: preload monaco (#728)
* refactor: remove unused "window.ElectronFiddle.contentChangeListeners

* refactor: remove unused "window.ElectronFiddle.editors"

* refactor: add renderer/main.tsx

This now handles the App creation and sentry tests that were previously
executed as a side-effect of importing 'app', which is a little clumsy
as we have to import 'app' in the tests -- and we also imported it in
ambient.d.ts, which makes a fun dependency loop.

* refactor: preload monaco before instantiating App.

Simplify the loading path so that individual pieces don't have to
asynchronously try to load monaco separately. Right now the Outputs
and Editors components are doing it, and EditorMosiac will also need
monaco in an upcoming refactor.

* refactor: remove loadMonaco()

With all of the loading + theme logic removed from Editors.loadMonaco()
and Outputs.loadMonaco(), the functions aren't needed anymore.

* refactor: remove some newly-dead components code

We no longer need to test for monaco-is-loaded since it's now preloaded.

* refactor: in fetch-types, remove monaco checks

Monaco is now preloaded, so we don't need the defer-until-loaded code.

* refactor: remove type ambiguity in preload

The `foo = foo || {}` assignment worked around typescript's type checks
for the window.ElectronFiddle type. This commit creates the whole object
in one pass.

* refactor: remove unused Editor, Output properties

* refactor: make monaco state properties readonly

They're never reassigned, so mark as readonly for a little type safety.

* refactor: move App.monaco to ElectronFiddle.monaco

This works around a bootstrap issue where code would look for
`ElectronFiddle.app.monaco` before the app was done being created.

It also extracts the Monaco mock into a new class which is named,
surprise surprise, MonacoMock.