refactor: try to make editor-mosaic internals more readable
This is a subjective thing, so feedback is welcomed. I've added more
comments to make some of the trickier implicit things explicit.

Also I've moved some of the functions around and reorganized the file
a little to make the flow from resetLayout() -> set() -> addFile() ->
setEditorFromBackup() -> show() a little easier to follow.