Electron 15 will have alpha relases. Don't break.
Fix sorting so that the progression is nightly > alpha > beta > stable.
This is important for the drop-down list and for bisection.

Fix filtering by including alphas in ElectronReleaseChannel.beta.

We want sorting / filtering to work, but adding a separate alpha release
channel would be overkill since alphas are a one-off process for E15.