perf: prefer getKnownVersions() over getElectronVersions().
The first returns a Version[] and the other returns a RunnableVersion[].
RunnableVersions are more expensive to create because calculating the
state property requires checking the disk for the version's exec file.

isReleasedMajor() and getOldestSupportedVersion() only need Version[],
so this PR changes them to call getKnownVersions().

Also, I think the old code was that way because I was unclear on the
meaning of `getKnownVersions()` -- local versions are also 'known'.
This PR renames it as `getReleasedVersions()` for clarity.